+++ packages/patternfly-4/react-integration/demo-app-ts/src/components/demos/index.ts
@@ -3,6 +3,7 @@ export * from './AccordionDemo/AccordionDemo';
 export * from './ApplicationLauncherDemo/ApplicationLauncherDemo';
 export * from './ApplicationLauncherDemo/ApplicationLauncherFavoritesDemo';
 export * from './AlertDemo/AlertDemo';
+export * from './AlertGroupDemo/AlertGroupDemo';
 export * from './AreaChart/SparklineDemo';
 export * from './AreaChart/CyanAreaDemo';
 export * from './AreaChart/ColorAreaBottomLegendDemo';