+++ packages/patternfly-4/react-catalog-view-extension/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.2.4...@patternfly/react-catalog-view-extension@1.2.5) (2020-01-27)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [1.2.4](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.2.3...@patternfly/react-catalog-view-extension@1.2.4) (2020-01-27)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [1.2.3](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.2.2...@patternfly/react-catalog-view-extension@1.2.3) (2020-01-27)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
 ## [1.2.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.2.1...@patternfly/react-catalog-view-extension@1.2.2) (2020-01-27)
 
 **Note:** Version bump only for package @patternfly/react-catalog-view-extension