+++ packages/patternfly-4/react-core/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.134.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.134.1...@patternfly/react-core@3.134.2) (2020-01-27)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
+## [3.134.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.134.0...@patternfly/react-core@3.134.1) (2020-01-27)
+
+
+### Bug Fixes
+
+* **dropdown:** do not pass id to dropdown list item if is not set ([#3555](https://github.com/patternfly/patternfly-react/issues/3555)) ([06df46a](https://github.com/patternfly/patternfly-react/commit/06df46a845da13a981239c9f7393258112161acf))
+
+
+
+
+
+# [3.134.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.133.2...@patternfly/react-core@3.134.0) (2020-01-27)
+
+
+### Features
+
+* **AlertGroup:** Add alert group component ([#3531](https://github.com/patternfly/patternfly-react/issues/3531)) ([0a6806e](https://github.com/patternfly/patternfly-react/commit/0a6806e0f6be01a022e6ce2a9d578a6d8bd52554))
+
+
+
+
+
 ## [3.133.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.133.1...@patternfly/react-core@3.133.2) (2020-01-27)
 
 **Note:** Version bump only for package @patternfly/react-core