+++ packages/react-icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.14.39](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@3.14.38...@patternfly/react-icons@3.14.39) (2020-01-27)
+
+**Note:** Version bump only for package @patternfly/react-icons
+
+
+
+
+
 ## [3.14.38](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@3.14.37...@patternfly/react-icons@3.14.38) (2020-01-27)
 
 **Note:** Version bump only for package @patternfly/react-icons